pyproject.toml
[project]
name = "text2video"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "loguru>=0.7.3",
    "moviepy>=2.2.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
